requirements.txt
neo4j>=5.0,<6.0
fastapi>=0.110
uvicorn>=0.30
google-genai
requests>=2.31
pydantic>=2.0
python-dotenv>=1.0
